gerrit-verifications.yml: Remove FreeBSD_amd64 builds
In early September we asked on the public mailing list if there are
any users of the FreeBSD builds, and there was no response at all.

Let's disable the build testing on FreeBSD.  This will significantly
speed up our build testing, as well as pave the way for a more
comprehensive docker/containerization of build testing.

We're still extremely happy to merge any patches for support of
FreeBSD or other operating systems.  But the core Osmocom developers
will not perform related testing / porting.

Change-Id: I2c6d2a17c3cf9d8c78c3675995493e30cbc6be0d