jobs: gerrit: fix missing indents in 4 jobs
fixed in osmo-bsc, osmo-mgw, osmo-sgsn, cellmgr-ng:

Spanning a single shell command across several lines with backslashes in the
end breaks when the newlines are not preserved: the backslashes escape a
following space, which is joined to the following cmdline arg.

Add the leading less-indented comments that curiously lead to preserving the
newline characters in the cmd sections.

Change-Id: Icfd6cfb7ca4172795620e1d7ee60610db4f7226b