osmocom packages: Ensure PACKAGE_VERSION is set correctly
At the "autorecon -fi" stage, Osmocom programs either need the .git
directory peresent, or a .tarball-version file in order to determine
the exact source code version.

Normally, "make dist" exists exactly for this purpose: It runs
git-version-gen and saves the result to .tarball-version, and we then
include this file in the .tar.gz we generate.

However, as the nightly paackaging scripts use git-buildpackage, it
bypasses the "make dist" logic and hence we need to

1) manually generate the .tarball-version file
2) copy it over to the directory specified as --git-export-dir

This way, the .tarball-version is inside the tar.xz generated by gbp,
and autoreconf then has something to use as PACKAGE_VERSION.

This commit fixes "UNKNOWN-dirty" in .pc files of libraries, as well
as in "show version" commands on the vty.

Closes: OS#3449
Change-Id: I76e3713f0b01a6110091ff90e8e53aa79533c374