osmocom-latest-packages.sh: Refactor to support building limesuite
This way the latest script is a lot more similar to the nightly one, and
easier to maintain and expand with new features.

checkout and build steps are split because once we have a new
osmo-trx release, we'll need the create_osmo_trx_debian8_jessie trick to
build it.

In the future we can do further steps to have a common function lib
between latest and nightly scripts.

Change-Id: I786c6f4ad4b4e43d1692c1588d2ad2194d0b25a4