jobs: osmo-gsm-teser: Run only once nightly
Current entire default testsuite takes almost 7 hours to run, so let's
run it once nightly and be done with it.
This way, during office time if we want to test something on the prod
setup (due to hw only avaialble there) we can do it without having to
wait, disable the job hours in advance or having to kill all processes
and clean osmo-gsm-tester state.

Change-Id: I138bf50d6d066d2da4cd97f25880628b93ddc237