osmocom-nightly-packages: build osmo-remsim
As of Change-Id Id5044b1835190edc948952d207a5196a18669eb1, osmo-remsim
contains Debian packaging information and hence we can build it as part
of our nightly feeds.

We're not yet enabling 'latest' builds as we need to wait for the next
libosmo-abis release for that (requires the IPA keepalive FSM).

Change-Id: Ic367dc95f46cece4bd769061af860fd82b4bd2e9