build-old-tags: remove LD_LIBRARY_PATH
Do not set LD_LIBRARY_PATH during builds, as this causes testsuites to
use the wrong libraries.

This bug appeared with libosmocore, it gets built for master first, and
then an old version like 1.2.0. When using LD_LIBRARY_PATH, the tests
during the 1.2.0 build are executed against the libosmo*.so from
master, which causes a few tests to fail.

Change-Id: I0bfb57e418b91c298337b9426448fbcfd7bf32e6