repo-install-test: update jenkins job
Rename osmocom-debian-install.yml to repo-install-test.yml to get debian
out of the name. Extend it with a new distro parameter and update the
description. Adjust the shell section to run the script from its new
location (in osmo-ci.git, not docker-playground.git). Turn it into a
matrix job, so we can have two parameters (distro, feed) for each job.

Related: OS#4563
Change-Id: I777098f19d75f7efbd68b837ccdcd83309429c39