ansible/gsm-tester: Install patchelf 0.11 on PATH
We used to install patchelf 0.10 in a specific path to only use it for
some binaries which had bugs with patchelf 0.9 (and other binaries had
bugs with 0.10, so we couldn't use 0.10 for all binaries and hence put
it in binary).

However, 0.11 is fixing both issues so we can use it in all cases, hence
we can put it in PATH (/usr/local/bin).

Related: OS#4389
Change-Id: Ica30807ecab7b60daee31fe7824b7c848d20c904