OBS: add %H%M to DT
Make sure that each time osmocom-nightly-packages.sh and
osmocom-next-packages.sh runs, we get a different date appended to the
versions, even if it runs twice on the same day (e.g. because the
jenkins job was triggered manually).

This is in preparation to let all packages depend on a specific version
(with that date) of the conflicting dummy package.

DT in osmocom-latest-packages.sh is adjusted for consistency (though it
is not appended to the package versions, only used in the commit message
when pushing the latest packages).

Related: OS#4733
Change-Id: I7f08c694a549f1b3dd938a68e05082f2c31fdb92