OBS: make rpmlint shlib-fixed-dependency non-fatal
Create a rpmlint config, which makes the shlib-fixed-dependency check
non-fatal, as it caused builds for openSUSE_Leap_15.2 to fail. The check
is supposed to warn about libraries depending on specific versions of
other packages. However, for the nightly and next packages, this is
exactly what we want to do to ensure that users will always upgrade all
Osmocom packages to the builds from a specific day, and not mix them.

Messages from the check:
[  307s] libosmocodec0.armv7hl: E: shlib-fixed-dependency (Badness: 440) osmocom-nightly = 1.0.0.202101181006
...
[  307s] libosmovty4.armv7hl: E: shlib-fixed-dependency (Badness: 440) osmocom-nightly = 1.0.0.202101181006
[  307s] Your shared library package requires a fixed version of another package. The
[  307s] intention of the Shared Library Policy is to allow parallel installation of
[  307s] multiple versions of the same shared library, hard dependencies likely make
[  307s] that impossible. Please remove this dependency and instead move it to the
[  307s] runtime uses of your library.
[  307s]
[  307s] (none): E: badness 3960 exceeds threshold 1000, aborting.

Related: OS#4733
Related: https://en.opensuse.org/openSUSE:Packaging_checks#Disarming_Fatal_Errors
Change-Id: I560b4adf80b5785d396a17afefa590559ad5ca5a