lint: checkpatch.pl: break exclude loop early
When running the linter on a patch with lots of files changed, and lots
of exclude files, this makes a big difference. On my machine, running an
osmo-iuh patch with 1580 files changed, and the high amount of asn1c
related excludes, the time for linting is reduced from 50s to 25s.

This should be acceptable, since typically we change only few files.

Change-Id: I6fb41dec25ecc1f2df7242ae041a8685a696c3fd