jobs/ttcn3: disable TTCN3-centos-remsim
It seems that TTCN3-centos-remsim-test was disabled manually via the web
UI and only got enabled recently after rolling out unrelated changes in
jobs/ttcn3-testsuites.yml. The tests do not run properly with the centos
version and it seems that nobody is using it, so disable them again for
now.

Change-Id: I4885333d63944af3510fdf0519233989808944ff