jobs: Automatically push ttcn3-bts-tests to osmocom registry
This image used to be pushed to the sysmocom registry, before the
osmocom one existed. Let's make it available in the osmocom one now that
we have it instead.

The new job triggers upon conclusion of the
registry-rebuild-upload-titan job, since anyway the ttcn3-bts-tests is
based on that image and hence requires it to be built prior to it.

Related: SYS#5765
Change-Id: I65e4e5a9e539de5bfd881e08d7c5842f2266ecee