scripts/OBS latest: fix checkout with dir in path
osmo-python-tests was added to latest just like it was in nightly, by
adding the following two lines to build_osmocom():

  checkout python/osmo-python-tests
  ...
  build osmo-python-tests

But the build fails, because the "checkout" function clones into a
"python/osmo-python-tests" directory in latest, and not into a
"osmo-python-tests" directory like in nightly. Make it behave the same
by adjusting the "checkout" function in nightly.

(This code has historically grown and should be refactored imho, but
let's fix this bug first.)

Fix for:
  ====> Building osmo-python-tests
  + cd /home/osmocom-build/jenkins/workspace/Osmocom_OBS_latest_obs.osmocom.org/osmo-python-tests
  ./scripts/osmocom-latest-packages.sh: 99: cd: can't cd to .../osmo-python-tests

Change-Id: I97074ee953cadac476fe2e4b686689c26bac1ea1