docker-rebuild: build erlang img on x86_64 only
Build debian-bullseye-erlang on x86_64 nodes only, as the image doesn't
get used on nodes with other arches. The image is only used by
gerrit-verifications jobs for osmo_dia2gsum and osmo_gsup, which are
restriected to run on these nodes:
https://jenkins.osmocom.org/jenkins/label/osmocom-gerrit-debian9/

Remove the check and comments for not building for debian 10 x32, as
that node has been replaced with a debian 11 one.

Related: OS#5453
Change-Id: I5636dfc9e06a44b096f9dc55528d677ea2dfc352