obs-mirror.sh: exclude dotfiles
There's a .~tmp~ file on the remote that currently causes the rsync
command to fail. Ignore all files starting with a dot.

Fix for:
  rsync: [sender] opendir "/opensuse/repositories/network:/osmocom:/nightly/Raspbian_11/.~tmp~" (in opensuse-full-really-everything-including-repositories) failed: Permission denied (13)

It's time to update the script to sync the obs.osmocom.org packages
instead of the ones from build.opensuse.org, I'll do that in the next
patch. Nevertheless, I think it makes sense to add this exclude argument
in case OBS (our instance) also creates temporary files like this.

Change-Id: Iacee2ea4acada60d8de960592fb828911a55781e