jobs: don't set node where it doesn't make sense
Remove the node attribute as these can run on any node. I've configured
the three deb11 nodes with "Usage: Use this node as much as possible"
instead of previously "Only build jobs with label expression matching
this node", so simply not specifying a node is now possible. Previously
it wouldn't have any node to run on as all of them were configured that
way.

Related: OS#5793
Change-Id: I2d67daf1cc7930649d097af23c18990cc3ae55fd